requirements.txt
# Core dependencies
numpy>=1.20.0,<1.27
pandas>=1.3.0,<2.1
scikit-learn>=1.0.0,<1.4
matplotlib>=3.4.0,<3.8
seaborn>=0.11.0,<0.13
plotly>=5.3.0,<5.20

# FastAPI and dependencies
fastapi>=0.68.0,<0.110
uvicorn>=0.15.0,<0.30
python-multipart>=0.0.5,<0.0.7
pydantic>=1.8.0,<2.0
requests>=2.26.0,<3.0
python-dotenv>=0.19.0,<1.0

# GitHub API
PyGithub>=1.55.0,<2.0

# GitPython Package
GitPython>=3.1.0,<4.0

# Machine learning and explainability
xgboost>=1.5.0,<2.0
shap>=0.40.0,<0.45
lime>=0.2.0,<0.3

# NLP
nltk>=3.6.0,<4.0
spacy>=3.1.0,<3.7
sentence-transformers>=2.2.0,<3.0

# PDF, HTML, and DOCX parsing
pymupdf>=1.18.0,<1.23
beautifulsoup4>=4.9.0,<5.0
python-docx>=0.8.11,<0.9

# Web scraping and automation
playwright>=1.20.0,<1.43

# Workflow orchestration
schedule>=1.1.0,<2.0

# Code analysis
astroid>=2.8.0,<3.0
radon>=5.1.0,<6.0

# Testing
pytest>=6.2.5,<8.0
pytest-cov>=2.12.0,<5.0

# Documentation
Sphinx>=4.2.0,<6.0
sphinx-rtd-theme>=1.0.0,<2.0

# Development tools
black>=21.9b0,<24.0
isort>=5.9.0,<6.0
flake8>=3.9.0,<7.0
mypy>=0.910,<1.10

# Some packages (like spacy) can be installed separately if you're using Conda:
# conda install -c conda-forge spacy
# python -m spacy download en_core_web_lg

# Playwright requires a separate install step for browsers:
# playwright install